week 3/medical_assistant/requirements.txt
streamlit
langchain
pillow
pymupdf
deep-translator
openai
ollama
python-dotenv